tx/ducktape: move authorization test to a separate file
